changed styling
styling login en registreer
en refactor client side validation
