+++ ClientApp/src/components/pages/authenticatie/RegistreerPage.js
@@ -1,4 +1,4 @@
-import { Alert, Button, TextField, Typography } from "@mui/material";
+import { Alert, Button, TextField, Typography, Card, Box } from "@mui/material";
 import { useRef } from "react";
 import { Link } from "react-router-dom";
 import { useState } from "react";
@@ -10,29 +10,10 @@ export default function RegistreerPage() {
 
     const onSubmit = (e) => {
         e.preventDefault();
-        const formData = form.current;
         setError(null);
+        const formData = form.current;
 
         // alle velden in een array zetten
-        const dataArray = [
-            { name: "voornaam", value: formData["voornaam"].value },
-            { name: "email", value: formData["email"].value },
-            { name: "wachtwoord", value: formData["wachtwoord"].value },
-            { name: "herhaalWachtwoord", value: formData["herhaalWachtwoord"].value },
-        ];
-
-        // checken of er lege velden zijn
-        const emptyFields = dataArray.filter((field) => field.value === "");
-        if (emptyFields.length > 1) {
-            setError(
-                "Vul de volgende velden in: " +
-                    emptyFields.map((field) => field.name).join(", ")
-            );
-            return;
-        } else if (emptyFields.length === 1) {
-            setError("Vul het volgende veld in: " + emptyFields[0].name);
-            return;
-        }
 
         // checken of wachtwoorden overeenkomen
         if (formData["wachtwoord"].value !== formData["herhaalWachtwoord"].value) {
@@ -59,63 +40,71 @@ export default function RegistreerPage() {
                     justifyContent: "center",
                     alignItems: "center",
                     pt: 10,
+                    backgroundColor: "#f5f5f5",
                 }}
             >
-                <div
-                    style={{
-                        width: 300,
-                        display: "flex",
-                        flexDirection: "column",
-                        gap: 16,
-                    }}
-                >
-                    <div>
-                        <Typography variant="h5" component="h1">
-                            Registreer
-                        </Typography>
-                        <Typography variant="body1" component="p">
-                            Registreer om door te gaan.
-                        </Typography>
-                    </div>
-                    {error && <Alert severity="error">{error}</Alert>}
-                    <form
-                        ref={form}
-                        style={{
+                <Card variant="outlined">
+                    <Box
+                        sx={{
                             width: 300,
                             display: "flex",
                             flexDirection: "column",
-                            gap: 16,
+                            gap: 3,
+                            m: 3,
                         }}
                     >
-                        <TextField
-                            name="voornaam"
-                            type="text"
-                            placeholder="Voornaam"
-                            label="Voornaam"
-                        />
-                        <TextField
-                            name="email"
-                            type="email"
-                            placeholder="name@mail.com"
-                            label="Email"
-                        />
-                        <TextField name="wachtwoord" type="password" label="Wachtwoord" />
-                        <TextField
-                            name="herhaalWachtwoord"
-                            type="password"
-                            label="Herhaal wachtwoord"
-                        />
-                    </form>
-                    <Button variant="contained" color="primary" onClick={onSubmit}>
-                        Registreer
-                    </Button>
-                    <Typography fontSize="sm" sx={{ alignSelf: "center" }}>
-                        Al een account?{" "}
-                        <Link style={{ color: "#1565c0" }} to="/login">
-                            Log in
-                        </Link>
-                    </Typography>
-                </div>
+                        <div>
+                            <Typography variant="h5" component="h1">
+                                Registreer
+                            </Typography>
+                            <Typography variant="body1" component="p">
+                                Registreer om door te gaan.
+                            </Typography>
+                        </div>
+                        {error && <Alert severity="error">{error}</Alert>}
+                        <form
+                            ref={form}
+                            style={{
+                                width: 300,
+                                display: "flex",
+                                flexDirection: "column",
+                                gap: 16,
+                            }}
+                        >
+                            <TextField
+                                name="voornaam"
+                                type="text"
+                                placeholder="Voornaam"
+                                label="Voornaam"
+                            />
+                            <TextField
+                                name="email"
+                                type="email"
+                                placeholder="name@mail.com"
+                                label="Email"
+                            />
+                            <TextField
+                                name="wachtwoord"
+                                type="password"
+                                label="Wachtwoord"
+                            />
+                            <TextField
+                                name="herhaalWachtwoord"
+                                type="password"
+                                label="Herhaal wachtwoord"
+                            />
+                        </form>
+                        <Button variant="contained" color="primary" onClick={onSubmit}>
+                            Registreer
+                        </Button>
+                        <Typography fontSize="sm" sx={{ alignSelf: "center" }}>
+                            Al een account?{" "}
+                            <Link style={{ color: "#1565c0" }} to="/login">
+                                Log in
+                            </Link>
+                        </Typography>
+                    </Box>
+                </Card>
             </div>
             <Footer />
         </div>
