+++ ClientApp/src/components/pages/home/HomePage.js
@@ -4,7 +4,7 @@ import Footer from "../../footer/Footer";
 export default function HomePage(props) {
     return (
         <div>
-            <Container maxWidth="xl">
+            <Container maxWidth="xl" sx={{ minHeight: "100vh" }}>
                 <h1>Home</h1>
             </Container>
             <Footer />
